proportions refactor: uppdate product and sale item DTOs

+++ backend/src/main/java/com/superpos/dto/SaleItemResponse.java
@@ -10,7 +10,7 @@ import java.math.BigDecimal;
 public class SaleItemResponse {
 
     private String productName;
-    private int quantity;
+    private BigDecimal quantity;
     private BigDecimal unitPrice;
     private BigDecimal subtotal;
